requirements.txt
Flask==2.1.1
opencv-python-headless==4.5.3.56
Pillow==8.3.1
pytesseract==0.3.8
numpy==1.21.2
difflib==3.7.4
pymongo==3.12.0
